app/requirements.txt
flask
supabase
python-dotenv
requests
app
gunicorn
flask-cors